+++ MuckFoods/CreateRecipesPatch.cs
@@ -5,7 +5,6 @@ using HarmonyLib;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
-
 public class CreateRecipesPatch
 {
 	public static InventoryItem[] NewFoods = new InventoryItem[0];
@@ -18,6 +17,7 @@ public class CreateRecipesPatch
 		var dough = ItemManager.Instance.allItems[47];
 		var flax = ItemManager.Instance.allItems[49];
 		var rawMeat = ItemManager.Instance.allItems[50];
+
 		var gulponShroom = ItemManager.Instance.allItems[51];
 		var ligonShroom = ItemManager.Instance.allItems[52];
 		var slurbonShroom = ItemManager.Instance.allItems[53];
@@ -34,25 +34,56 @@ public class CreateRecipesPatch
 		var weirdSoup = ItemManager.Instance.allItems[63];
 		var yellowSoup = ItemManager.Instance.allItems[64];
 
-		// Adjust vanilla foods so things don't get too insane
+		// Just make sure inedible items don't add any stats to foods
+		SetStats(bowl, 0f, 0f, 0f);
+		SetStats(flax, 0f, 0f, 0f);
+
+		// We adjust vanilla foods so things don't get too insane
 		// Meat soup and cooked meat are super easy to get, so they are nerfed
 		if (Plugin.RebalanceFoods)
 		{
-			apple.hunger = 5f;
-			rawMeat.hunger = 5f;
+			SetStats(apple, 4f, 4f, 4f);
+			SetStats(rawMeat, 6f, 6f, 3f);
 			gulponShroom.heal = 10f;
 			ligonShroom.hunger = 10f;
 			sugonShroom.stamina = 10f;
 			SetStats(slurbonShroom, 10f, 10f, 10f);
-			SetStats(cookedMeat, 15f, 15f, 5f);
-			SetStats(meatSoup, 20f, 20f, 10f);
-			SetStats(applePie, 35f, 35f, 35f);
-			SetStats(meatPie, 55f, 55f, 40f);
-			SetStats(purpleSoup, 10f, 10f, 30f);
-			SetStats(redSoup, 30f, 10f, 10f);
+
+			SetStats(redSoup, 25f, 10f, 10f);
+			SetStats(yellowSoup, 10f, 25f, 10f);
+			SetStats(purpleSoup, 10f, 10f, 25f);
 			SetStats(weirdSoup, 30f, 30f, 30f);
-			SetStats(yellowSoup, 10f, 30f, 10f);
+
+			Restoration.FromIngredients(rawMeat).Apply(cookedMeat);
+			Restoration.FromIngredients(cookedMeat, bowl).Apply(meatSoup);
+			Restoration.FromIngredients(apple, bread, bowl).Apply(applePie);
+			Restoration.FromIngredients(cookedMeat, bowl, bread).Apply(meatPie);
 		}
+
+		var bakedApple = CreateFood(apple, Resources.AppleBakedSprite, "Baked Apple", "soft and sweet, but looks familiar somehow...", CraftRequirements(apple), apple);
+		var toastyGulponShroom = CreateFood(gulponShroom, Resources.ShroomRedToastySprite, "Toasted Gulpon Shroom", "nice and hot!", CraftRequirements(gulponShroom), gulponShroom);
+		var toastyLigonShroom = CreateFood(ligonShroom, Resources.ShroomYellowToastySprite, "Toasted Ligon Shroom", "tasty and filling", CraftRequirements(ligonShroom), ligonShroom);
+		var toastySugonShroom = CreateFood(sugonShroom, Resources.ShroomPurpleToastySprite, "Toasted Sugon Shroom", "keeps you jumping at your best", CraftRequirements(sugonShroom), sugonShroom);
+		var toastySlurbonShroom = CreateFood(slurbonShroom, Resources.ShroomSusToastySprite, "Toasted Slurbon Shroom", "still kinda sus...", CraftRequirements(slurbonShroom), slurbonShroom);
+		bakedApple.processTime = cookedMeat.processTime;
+		toastyGulponShroom.processTime = cookedMeat.processTime;
+		toastyLigonShroom.processTime = cookedMeat.processTime;
+		toastySugonShroom.processTime = cookedMeat.processTime;
+		toastySlurbonShroom.processTime = cookedMeat.processTime;
+
+		List<InventoryItem> newItems = new()
+		{
+			bakedApple,
+			toastyGulponShroom,
+			toastyLigonShroom,
+			toastySugonShroom,
+			toastySlurbonShroom,
+			CreateFood(applePie, applePie.sprite, "Pie Pie", "it's pie flavoured", CraftRequirements(dough, bowl), bread, bowl),
+			CreateFood(bread, Resources.FlaxseedBreadSprite, "Flaxseed Bread", "what else can you do with flax?", CraftRequirements(dough, flax), bread, flax),
+			CreateFood(meatSoup, Resources.AppleMeatStewSprite, "Apple Meat Stew", "yummy", CraftRequirements(apple, rawMeat, bowl), apple, cookedMeat, bowl),
+			CreateFood(applePie, Resources.PieAppleMeatSprite, "Apple Meat Pie", "dessert and dinner in one", CraftRequirements(apple, rawMeat, dough, bowl), apple, cookedMeat, bread, bowl),
+			CreateFood(weirdSoup, Resources.LessWeirdSoupSprite, "Less Weird Soup", "food for a fun guy", CraftRequirements(gulponShroom, ligonShroom, sugonShroom, bowl), gulponShroom, ligonShroom, sugonShroom, bowl)
+		};
 
 		var shrooms = new (InventoryItem room, string name, string colour, InventoryItem soup)[]
 		{
@@ -61,36 +92,61 @@ public class CreateRecipesPatch
 			(sugonShroom, "Sugon", "Purple", purpleSoup),
 			(slurbonShroom, "Slurbon", "Weird", weirdSoup),
 		};
-
-		// The way we have made things work is basically every food restores the same as if you cooked everything and ate it, plus 5 to everything.
-		// For meat, we consider meat soup, instead of cooked meat.
-		List<InventoryItem> newItems = new()
+		var descrShroomPie = new DescrSprite[]
 		{
-			CreateFood(meatSoup, "Applesauce", "perfect for weak teeth", CraftRequirements(apple, bowl), apple),
-			CreateFood(applePie, "Pie Pie", "it's pie flavoured", CraftRequirements(dough, bowl), bread),
-			CreateFood(bread, "Flaxseed Bread", "what else can one do with flax?", CraftRequirements(dough, flax), bread),
-			CreateFood(meatSoup, "Apple Meat Stew", "yummy", CraftRequirements(apple, rawMeat, bowl), apple, meatSoup),
-			CreateFood(applePie, "Apple Meat Pie", "dessert and dinner in one", CraftRequirements(apple, rawMeat, dough, bowl), apple, rawMeat, bread),
-			CreateFood(weirdSoup, "Less Weird Soup", "food for a fun guy", CraftRequirements(gulponShroom, ligonShroom, sugonShroom, bowl), redSoup, yellowSoup, purpleSoup),
+			new("spicy pie", Resources.PieRedSprite),
+			new("pie with shroom", Resources.PieYellowSprite),
+			new("pie in the sky", Resources.PiePurpleSprite),
+			new("mystery pie", Resources.PieSusSprite),
 		};
-
+		var descrShroomApplePie = new DescrSprite[]
+		{
+			new("pretty healthy", Resources.PieRedAppleSprite),
+			new("pie with apple and shroom", Resources.PieYellowAppleSprite),
+			new("pretty fast pie", Resources.PiePurpleAppleSprite),
+			new("who needs meat anyway?", Resources.PieSusAppleSprite),
+		};
+		var descrShroomMeatPie = new DescrSprite[]
+		{
+			new("don't die pie", Resources.PieRedMeatSprite),
+			new("could feed an army (or just you)", Resources.PieYellowMeatSprite),
+			new("fills you with energy", Resources.PiePurpleMeatSprite),
+			new("sus, savoury, scrumptious", Resources.PieSusMeatSprite),
+		};
+		var descrShroomAppleStew = new DescrSprite[]
+		{
+			new("red and vegan friendly", Resources.SoupRedAppleSprite),
+			new("it's sweet and savoury", Resources.SoupYellowAppleSprite),
+			new("better than nothing ig", Resources.SoupPurpleAppleSprite),
+			new("lv 5 vegans only", Resources.SoupSusAppleSprite),
+		};
+		var descrShroomMeatStew = new DescrSprite[]
+		{
+			new("savoury stew", Resources.SoupRedMeatSprite),
+			new("tender meat and shroom", Resources.SoupYellowMeatSprite),
+			new("tasty stew for when you're tired", Resources.SoupPurpleMeatSprite),
+			new("i think the meat became rainbows", Resources.SoupSusMeatSprite),
+		};
+		var descrShroomAppleMeatStew = new DescrSprite[]
+		{
+			new("reddest food ever made", Resources.SoupRedAppleMeatSprite),
+			new("a real hunter-gatherer's meal", Resources.SoupYellowAppleMeatSprite),
+			new("keeps you going", Resources.SoupPurpleAppleMeatSprite),
+			new("the heartiest stew in the land", Resources.SoupSusAppleMeatSprite),
+		};
 		int i = 0;
-		var descrShroomPie = new string[] { "spicy pie", "pie with shroom", "pie in the sky", "mystery pie" };
-		var descrShroomMeatPie = new string[] { "don't die pie", "could feed an army (or just you)", "fills you with energy", "sus, savoury, scrumptious" };
-		var descrShroomApplePie = new string[] { "pretty healthy", "pie with apple and shroom", "pretty fast pie", "who needs meat anyway?" };
-		var descrShroomAppleStew = new string[] { "red and vegan friendly", "it's sweet and savoury", "better than nothing ig", "lv 5 vegans only" };
-		var descrShroomMeatStew = new string[] { "savoury stew", "tender meat and shroom", "tasty stew for when you're tired", "i think the meat became rainbows" };
-		var descrShroomAppleMeatStew = new string[] { "reddest food ever made", "a real hunter-gatherer's meal", "keeps you going", "the heartiest stew in the land" };
 		foreach (var mush in shrooms)
 		{
-			newItems.Add(CreateFood(applePie, mush.name + " Shroom Pie", descrShroomPie[i], CraftRequirements(dough, mush.room, bowl), bread, mush.soup));
-			newItems.Add(CreateFood(applePie, string.Concat("Apple and ", mush.name, " Shroom Pie"), descrShroomMeatPie[i], CraftRequirements(dough, apple, mush.room, bowl), bread, apple, mush.soup));
-			newItems.Add(CreateFood(applePie, string.Concat("Meat and ", mush.name, " Shroom Pie"), descrShroomApplePie[i], CraftRequirements(dough, rawMeat, mush.room, bowl), bread, meatSoup, mush.soup));
-			newItems.Add(CreateFood(mush.soup, string.Concat("Apple and ", mush.name, " Shroom Stew"), descrShroomAppleStew[i], CraftRequirements(apple, mush.room, bowl), apple, mush.soup));
-			newItems.Add(CreateFood(mush.soup, string.Concat("Meat and ", mush.name, " Shroom Stew"), descrShroomMeatStew[i], CraftRequirements(rawMeat, mush.room, bowl), meatSoup, mush.soup));
-			newItems.Add(CreateFood(mush.soup, mush.colour + " Gourmet Stew", descrShroomAppleMeatStew[i], CraftRequirements(rawMeat, apple, mush.room, bowl), meatSoup, apple, mush.soup));
+			newItems.Add(CreateFood(applePie, descrShroomPie[i].Sprite, mush.name + " Shroom Pie", descrShroomPie[i].Descr, CraftRequirements(dough, mush.room, bowl), bread, mush.soup, bowl));
+			newItems.Add(CreateFood(applePie, descrShroomApplePie[i].Sprite, string.Concat("Apple and ", mush.name, " Shroom Pie"), descrShroomApplePie[i].Descr, CraftRequirements(dough, apple, mush.room, bowl), bread, apple, mush.soup, bowl));
+			newItems.Add(CreateFood(applePie, descrShroomMeatPie[i].Sprite, string.Concat("Meat and ", mush.name, " Shroom Pie"), descrShroomMeatPie[i].Descr, CraftRequirements(dough, rawMeat, mush.room, bowl), bread, cookedMeat, mush.soup, bowl));
+			newItems.Add(CreateFood(mush.soup, descrShroomAppleStew[i].Sprite, string.Concat("Apple and ", mush.name, " Shroom Stew"), descrShroomAppleStew[i].Descr, CraftRequirements(apple, mush.room, bowl), apple, mush.soup, bowl));
+			newItems.Add(CreateFood(mush.soup, descrShroomMeatStew[i].Sprite, string.Concat("Meat and ", mush.name, " Shroom Stew"), descrShroomMeatStew[i].Descr, CraftRequirements(rawMeat, mush.room, bowl), cookedMeat, mush.soup, bowl));
+			newItems.Add(CreateFood(mush.soup, descrShroomAppleMeatStew[i].Sprite, mush.colour + " Gourmet Stew", descrShroomAppleMeatStew[i].Descr, CraftRequirements(rawMeat, apple, mush.room, bowl), cookedMeat, apple, mush.soup, bowl));
 			++i;
 		}
+		applePie.sprite = Resources.PieAppleSprite;
+		meatPie.sprite = Resources.PieMeatSprite;
 
 		// Now we assign IDs to all the newly created items
 		// The way IDs work is basically Muck has the array "allScriptableItems", and IDs are assigned based on index.
@@ -111,23 +167,12 @@ public class CreateRecipesPatch
 	{
 		CauldronUI.Instance.processableFood = CauldronUI.Instance.processableFood.Concat(NewFoods).ToArray();
 	}
-	private static InventoryItem CreateFood(InventoryItem baseItem, string name, string descr, InventoryItem.CraftRequirement[] requirements, params InventoryItem[] restoresTheSumOfThisStuff)
+	private static InventoryItem CreateFood(InventoryItem baseItem, Sprite sprite, string name, string descr, InventoryItem.CraftRequirement[] requirements, params InventoryItem[] restoresSumOf)
 	{
-		float heal = Plugin.HealthBonus[restoresTheSumOfThisStuff.Length - 1];
-		float hunger = Plugin.HungerBonus[restoresTheSumOfThisStuff.Length - 1];
-		float stamina = Plugin.StaminaBonus[restoresTheSumOfThisStuff.Length - 1];
-		foreach (var s in restoresTheSumOfThisStuff)
-		{
-			heal += s.heal;
-			hunger += s.hunger;
-			stamina += s.stamina;
-		}
-		InventoryItem food = Clone(baseItem);
-		food.name = name;
-		food.description = descr;
-		food.requirements = requirements;
-		SetStats(food, heal, hunger, stamina);
-		Plugin.Log.LogInfo(string.Concat("Added food \"", name, "\" which restores ", heal, " Health, ", hunger, " Hunger, ", stamina, " Stamina. Made from: ", string.Join(",", food.requirements.Select(x => x.item.name))));
+		Plugin.Log.LogInfo("Making: " + name);
+		Restoration resto = Restoration.FromIngredients(restoresSumOf);
+		InventoryItem food = Clone(baseItem, name, descr, sprite, requirements, resto);
+		Plugin.Log.LogInfo(string.Concat("Added food \"", name, "\" which restores ", resto.Heal, " Health, ", resto.Hunger, " Hunger, ", resto.Stamina, " Stamina. Made from: ", string.Join(",", food.requirements.Select(x => x.item.name))));
 		return food;
 	}
 	private static InventoryItem.CraftRequirement[] CraftRequirements(params InventoryItem[] items)
@@ -145,17 +190,26 @@ public class CreateRecipesPatch
 		food.hunger = hunger;
 		food.stamina = stamina;
 	}
-	private static InventoryItem Clone(InventoryItem item)
+	private static InventoryItem Clone(InventoryItem item, string name, string description, Sprite sprite, InventoryItem.CraftRequirement[] requirements, Restoration resto)
 	{
 		InventoryItem i = ScriptableObject.CreateInstance<InventoryItem>();
+		i.material = item.material;
+		i.sprite = sprite;
+		i.name = name;
+		i.description = description;
+		i.requirements = requirements;
+
+		resto.Apply(i);
+
+		// Not necessary
+		//i.id = item.id;
+		//i.heal = item.heal;
+		//i.hunger = item.hunger;
+		//i.stamina = item.stamina;
+
 		i.important = item.important;
-		i.id = item.id;
-		i.name = item.name;
-		i.description = item.description;
 		i.type = item.type;
 		i.tier = item.tier;
-		i.sprite = item.sprite;
-		i.material = item.material;
 		i.mesh = item.mesh;
 		i.rotationOffset = item.rotationOffset;
 		i.positionOffset = item.positionOffset;
@@ -179,9 +233,6 @@ public class CreateRecipesPatch
 		i.processType = item.processType;
 		i.processedItem = item.processedItem;
 		i.processTime = item.processTime;
-		i.heal = item.heal;
-		i.hunger = item.hunger;
-		i.stamina = item.stamina;
 		i.armor = item.armor;
 		i.swingFx = item.swingFx;
 		i.bowComponent = item.bowComponent;
@@ -192,7 +243,6 @@ public class CreateRecipesPatch
 		i.attackTypes = item.attackTypes;
 		i.hideFlags = item.hideFlags;
 		i.craftAmount = item.amount;
-		i.requirements = item.requirements;
 		return i;
 	}
 }