+++ MuckFoods/CHANGELOG.md
@@ -6,3 +6,11 @@
 
 # v1.2.0
 - Added Flaxseed Bread, made from Dough and Flax
+
+# v1.3.0
+- Sprites for all the new foods (although not very good)
+- Changed Applesauce to Baked Apple, and it no longer requires a bowl
+- Added recipes for toasting all kinds of Shrooms
+- Food bonuses depend on the number of any type of ingredient, not just edible ones
+- Flat bonus added to cooked food
+- Apple and Raw Meat hunger restoration very slightly adjusted